refactor: rename forward to forward_req to prep for forward_call
